added progress updating for razipper step

+++ lib/razipper.rb
@@ -7,7 +7,6 @@ require_relative 'files_uploader'
 
 class Razipper
   include ConnectionValidation
-  include FilesUploader
 
   attr_reader :list_of_urls, :archive_folder, :archive_path, :archive_name
 
@@ -15,9 +14,13 @@ class Razipper
     @list_of_urls = list_of_urls.map { |url| URI(url) }
   end
 
-  def zip(name: 'archive', limit_files_size: 100 * 1024 * 1024)
-    easy_validate!(list_of_urls: list_of_urls, max_size: limit_files_size)
-    responses = grub_files_by_urls(list_of_urls, limit_files_size)
+  def zip(name: 'archive', limit_files_size: 100 * 1024 * 1024, progress_service: nil)
+    easy_validate!(bit_of_files: bit_of_files, max_size: limit_files_size)
+    progress_service.bits_of_files = bit_of_files if progress_service
+
+    files_uploader = FilesUploader.new(list_of_urls, limit_files_size)
+    responses = files_uploader.grub_files_by_urls(progress_service)
+
     @archive_path = generate_uniq_tmp_folder + '/' + name + '.zip'
     @archive_name = name
 
@@ -26,17 +29,29 @@ class Razipper
         file_name = "#{i + 1}_" << grub_file_name_from_path(url.path)
 
         zio.put_next_entry(file_name)
-        zio.print(responses.delete(url))
+        zio.print(responses.delete(url.to_s))
       end
     end
     archive_path
   end
 
   def remove_zip
+    return unless archive_folder
     FileUtils.remove_dir(archive_folder) if File.directory?(archive_folder)
   end
 
   private
+
+  def bit_of_files
+    hash = {}
+    list_of_urls.each do |url|
+      http = Net::HTTP.new(url.host, url.port)
+      http.use_ssl = (url.scheme == "https")
+
+      hash[url.to_s] = http.request_head(url)['content-length'].to_i
+    end
+    hash
+  end
 
   def generate_uniq_tmp_folder
     @archive_folder = FileUtils.mkdir_p("#{Dir.pwd}/tmp/#{SecureRandom.hex(5)}")[0]
